Create index.html
vanilla items and Tg / Sf don't work, deal with it. 

Tr isn't the exact same as $1 and $2 also I turn on Tg for $3, deal with it. 

S! isn't turned off for anything except S2 as is the same in the implementation of the current flags, but be aware of that fact. 

E2 and E are both changed to just Er, can't change that so complain to Board

There is no F flag equivalent, so both F2 and F are changed to just F